backend/requirements.txt
uvicorn
fastapi
requests
python-dotenv
httpx